move l1 commiter state diff build to common
