+++ crates/vm/constants.rs
@@ -12,20 +12,6 @@ pub static WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS: LazyLock<Address> =
 pub static CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS: LazyLock<Address> =
     LazyLock::new(|| Address::from_str("0000BBdDc7CE488642fb579F8B00f3a590007251").unwrap());
 
-// transactions_root(H256) + receipts_root(H256) + parent_hash(H256) + gas_limit(u64) + gas_used(u64) + timestamp(u64)
-// block_number(u64) + base_fee_per_gas(u64)
-// 32bytes + 32bytes + 32bytes + 8bytes + 8bytes + 8bytes + 8bytes + 8bytes
-// TODO: this should be changed to follow this issue https://github.com/lambdaclass/ethrex/issues/2655
-pub const LAST_HEADER_FIELDS_SIZE: usize = 136;
-
-// address(H160) + amount(U256) + tx_hash(H256).
-// 20bytes + 32bytes + 32bytes.
-pub const L2_WITHDRAWAL_SIZE: usize = 84;
-
-// address(H160) + amount(U256).
-// 20bytes + 32bytes
-pub const L2_DEPOSIT_SIZE: usize = 52;
-
 pub static COMMON_BRIDGE_L2_ADDRESS: LazyLock<Address> = LazyLock::new(|| {
     Address::from_slice(&[
         0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
